LayoutTests/imported/w3c/web-platform-tests/css/selectors/selector-after-font-family-expected.txt


PASS :empty is a valid selector
